refactor: extract suggest-date
